+++ backend/src/main/java/com/salonflow/backend/service/CustomerService.java
@@ -1,7 +1,7 @@
 package com.salonflow.backend.service;
 
-import com.salonflow.backend.controller.dtos.CustomerCreateDTO;
-import com.salonflow.backend.controller.dtos.CustomerDTO;
+import com.salonflow.backend.controller.dtos.customer.CustomerCreateDTO;
+import com.salonflow.backend.controller.dtos.customer.CustomerDTO;
 import com.salonflow.backend.controller.dtos.response.CustomerResponseDTO;
 import com.salonflow.backend.domain.model.Customer;
