+++ backend/src/main/java/com/salonflow/backend/service/ScheduleService.java
@@ -1,6 +1,6 @@
 package com.salonflow.backend.service;
 
-import com.salonflow.backend.controller.dtos.ScheduleDTO;
+import com.salonflow.backend.controller.dtos.schedule.ScheduleDTO;
 import com.salonflow.backend.domain.model.Schedule;
 
 public interface ScheduleService {
