+++ backend/src/test/java/com/salonflow/backend/service/Impl/ScheduleServiceImplTest.java
@@ -1,7 +1,7 @@
 package com.salonflow.backend.service.Impl;
 
-import com.salonflow.backend.controller.dtos.CustomerCreateDTO;
-import com.salonflow.backend.controller.dtos.ScheduleDTO;
+import com.salonflow.backend.controller.dtos.customer.CustomerCreateDTO;
+import com.salonflow.backend.controller.dtos.schedule.ScheduleDTO;
 import com.salonflow.backend.domain.model.Customer;
 import com.salonflow.backend.domain.model.Professional;
 import com.salonflow.backend.domain.model.ProfessionalServices;
